update profile screen begun - Facebook SignIn process was refactored
